[#3318] Prelude Table refcount overflow for selection with > 65535 rows
Since each individual row reference the table, we can easily overflow the 16 bits
refcount.

+++ src/preludedb-sql.c
@@ -111,7 +111,7 @@ struct preludedb_sql_table {
         unsigned int row_count;
         unsigned int column_count;
 
-        uint16_t refcount;
+        uint32_t refcount;
         uint8_t done;
 };
